requirements.txt
streamlit==1.54.0
google-genai==1.55.0
pandas==2.3.3
requests==2.32.5
matplotlib==3.10.8
python-dotenv==1.2.1
web3==6.15.1
plotly==6.0.0
streamlit-autorefresh==0.0.1
